clean up crs in cost layer + rename distance -> layer + multilayer display
